Poistettu näytön orientaatio, sovelluksessa pystysuunta Oletusnäkymään kirjautuminen ja rekisteröityminen, sovelluksen tietoutta siirretty piiloisaan Matiaksen tekemät refaktoroinnit huomioitu Vähän siistitty user deletionii
