+++ frontend/package.json
@@ -22,7 +22,6 @@
     "emulator": "^0.1.0",
     "expo": "^52.0.4",
     "expo-location": "18.0.1",
-    "expo-screen-orientation": "8.0.0",
     "expo-status-bar": "2.0.0",
     "firebase": "^10.14.1",
     "he": "^1.2.0",
